refactor: move filesystem tools to src/tools/fs/ directory
