+++ src/tools/mod.ts
@@ -90,50 +90,9 @@ export function createTool<T extends z.ZodObject<z.ZodRawShape>>(options: {
   };
 }
 
-// Tool exports
-export { ReadFileTool } from "./ReadFileTool.ts";
-export { ListDirTool } from "./ListDirTool.ts";
-export { createEditFileTools } from "./EditFileTool.ts";
+// Filesystem tools
+export * from "./fs/mod.ts";
+
+// Other tools
 export { RunTerminalCmdTool } from "./RunTerminalCmdTool.ts";
-export { GrepSearchTool } from "./GrepSearchTool.ts";
-export { FileSearchTool } from "./FileSearchTool.ts";
-export { CopyFileTool, DeleteFileTool } from "./FileTools.ts";
 export { createImageTools } from "./ImageTools.ts";
-
-// Re-import for createFileSystemTools
-import { ReadFileTool } from "./ReadFileTool.ts";
-import { ListDirTool } from "./ListDirTool.ts";
-import { createEditFileTools } from "./EditFileTool.ts";
-import { GrepSearchTool } from "./GrepSearchTool.ts";
-import { FileSearchTool } from "./FileSearchTool.ts";
-import { CopyFileTool, DeleteFileTool } from "./FileTools.ts";
-
-/**
- * Creates all built-in filesystem tools for easy registration.
- *
- * @param options - Optional configuration for the filesystem tools
- * @param options.backupDir - Custom backup directory for edit tools.
- *  If not provided, defaults to {workspaceDataDir}/backup.
- * @returns An array of all filesystem tools ready for registration
- *
- * @example
- * ```ts
- * const agent = await createZypherAgent({
- *   modelProvider,
- *   tools: createFileSystemTools(),
- * });
- * ```
- */
-export function createFileSystemTools(
-  options?: { backupDir?: string },
-): Tool[] {
-  return [
-    ReadFileTool,
-    ListDirTool,
-    ...createEditFileTools(options?.backupDir),
-    GrepSearchTool,
-    FileSearchTool,
-    CopyFileTool,
-    DeleteFileTool,
-  ];
-}
